Add populate db tool
