+++ backend/tools/data/user.json
@@ -1,0 +1,10 @@
+[
+  {
+    "id": 1,
+    "email": "test@example.com",
+    "password_hash": "password123",
+    "active": true,
+    "vendor_id": 1,
+    "metadata": {}
+  }
+]
