--- backend/tools/test_project_metadata.json
@@ -1,16 +1,0 @@
-{
-  "brand_name": "test brand",
-  "description": "test description",
-  "gdrive_link": "test gdrive link",
-  "shipping_terms": "test terms",
-  "cost_model": {
-    "wastepicker_sell_price": 1.00,
-    "wastepicker_purchase_price": 1.00,
-    "wholesaler_sell_price": 1.00,
-    "price_buoyancy": 1.00,
-    "wholesaler_exworks_price": 1.00,
-    "wholesaler_shipping_price": 1.00,
-    "wholesaler_delivered_price": 1.00,
-    "pfc_transaction_fee": 1.00
-  }
-}
